+++ modules/actor/sheet/cpr-character-sheet.js
@@ -239,12 +239,18 @@ export default class CPRCharacterActorSheet extends CPRActorSheet {
    */
   async _selectRoles() {
     LOGGER.trace("_selectRoles | CPRCharacterActorSheet | Called.");
-    let formData = { actor: this.actor.getData().roleInfo, roles: CPR.roleList };
+    if (this.actor.data.filteredItems.role.length === 0) {
+      SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.characterSheet.bottomPane.role.noRolesWarning"));
+      return;
+    }
+    let formData = {
+      roles: this.actor.data.filteredItems.role,
+    };
     formData = await SelectRolePrompt.RenderPrompt(formData).catch((err) => LOGGER.debug(err));
     if (formData === undefined) {
       return;
     }
-    await this.actor.setRoles(formData);
+    this.actor.update({ "data.roleInfo.activeRole": formData.activeRole, "data.roleInfo.activeNetRole": formData.activeNetRole });
   }
 
   /**
@@ -369,7 +375,7 @@ export default class CPRCharacterActorSheet extends CPRActorSheet {
       if (!Number.isNaN(parseInt(event.target.value, 10))) {
         item.setWeaponAmmo(event.target.value);
       } else {
-        SystemUtils.DisplayMessage("error", SystemUtils.Localize("CPR.messages.amountNotNumber"));
+        SystemUtils.DisplayMessage("error", SystemUtils.Localize("CPR.amountnotnumber"));
       }
     }
     this._updateOwnedItem(item);
@@ -405,21 +411,27 @@ export default class CPRCharacterActorSheet extends CPRActorSheet {
    * @param {*} event - object with details of the event
    */
   _updateRoleAbility(event) {
-    LOGGER.trace("_updateRoleAbility | CPRCharacterActorSheet | Called.");
-    const role = $(event.currentTarget).attr("data-role-name");
-    const ability = $(event.currentTarget).attr("data-ability-name");
+    LOGGER.trace("ActorID _updateRoleAbility | CPRCharacterActorSheet | Called.");
+    const item = this._getOwnedItem(CPRActorSheet._getItemId(event));
+    const itemData = duplicate(item.data);
     const subskill = $(event.currentTarget).attr("data-subskill-name");
     const value = parseInt(event.target.value, 10);
-    const actorData = duplicate(this.actor.data);
-    if (hasProperty(actorData, "data.roleInfo")) {
-      const prop = getProperty(actorData, "data.roleInfo");
-      if (subskill) {
-        prop.roleskills[role].subSkills[subskill] = value;
-      } else {
-        prop.roleskills[role][ability] = value;
+    if (!Number.isNaN(value)) {
+      if (hasProperty(itemData, "data.rank")) {
+        if (subskill) {
+          const updateSubskill = itemData.data.abilities.filter((a) => a.name === subskill);
+          if (updateSubskill.length === 1) {
+            updateSubskill[0].rank = value;
+          } else {
+            SystemUtils.DisplayMessage("error", SystemUtils.Localize("CPR.messages.multipleAbilitiesWithTheSameName"));
+          }
+        } else {
+          itemData.data.rank = value;
+        }
+        item.update(itemData);
       }
-      setProperty(actorData, "data.roleInfo", prop);
-      this.actor.update(actorData);
+    } else {
+      SystemUtils.DisplayMessage("error", SystemUtils.Localize("CPR.messages.amountNotNumber"));
     }
   }