+++ modules/hooks/item.js
@@ -1,5 +1,6 @@
 /* global Hooks */
 import LOGGER from "../utils/cpr-logger.js";
+import Rules from "../utils/cpr-rules.js";
 import SystemUtils from "../utils/cpr-systemUtils.js";
 
 /**
@@ -7,6 +8,34 @@ import SystemUtils from "../utils/cpr-systemUtils.js";
  * we document them all for clarity's sake and to make future development/debugging easier.
  */
 const itemHooks = () => {
+  /**
+   * The updateItem Hook is provided by Foundry and triggered here. When an Item is updated, this hook is called
+   * right after. When an item is updated (specifically a role item) we check to see if a multiplier is set.
+   * If it is, we set values for the "sub-roles."
+   *
+   * @public
+   * @memberof hookEvents
+   * @param {Document} doc          The Item document which is requested for creation
+   * @param {object} updateData     A trimmed object with the data provided for creation
+   * @param {object} (unused)       Additional options which modify the creation request
+   * @param {string} (unused)       The ID of the requesting user, always game.user.id
+   */
+  Hooks.on("updateItem", (doc, updateData) => {
+    LOGGER.trace("updateItem | itemHooks | Called.");
+    if (updateData.data && updateData.data.abilities) {
+      const roleRank = doc.data.data.rank;
+      let subRolesValue = 0;
+      doc.data.data.abilities.forEach((a) => {
+        if (a.multiplier !== "--") {
+          subRolesValue += (a.rank * a.multiplier);
+        }
+      });
+      if (subRolesValue > roleRank) {
+        Rules.lawyer(false, "CPR.messages.invalidRoleData");
+      }
+    }
+  });
+
   /**
    * The preCreateItem Hook is provided by Foundry and triggered here. When an Item is created, this hook is called just
    * prior to creation. This hook overrides Foundry's default item images when items are created in the sidebar.
@@ -29,6 +58,37 @@ const itemHooks = () => {
     if ((typeof createData.img === "undefined") && doc.parent === null) {
       const itemImage = SystemUtils.GetDefaultImage("Item", createData.type);
       doc.data.update({ img: itemImage });
+    }
+  });
+
+  /**
+   * The deleteItem Hook is provided by Foundry and triggered here. When an Item is deleted, this hook is called during
+   * deletion.
+   *
+   * @public
+   * @memberof hookEvents
+   * @param {CPRItem} itemData            The pending document which is requested for creation
+   * @param {object} (unused)             Additional options which modify the creation request
+   * @param {string} (unused)               The ID of the requesting user, always game.user.id
+   */
+
+  Hooks.on("deleteItem", (itemData) => {
+    LOGGER.trace("deleteItem | itemHooks | Called.");
+    const actor = itemData.parent;
+    if (actor !== null) {
+      if (itemData.type === "role" && actor.data.data.roleInfo.activeRole === itemData.name) {
+        if (actor.data.filteredItems.role.length >= 1) {
+          const newActiveRole = actor.data.filteredItems.role
+            .sort((a, b) => (a.data.name > b.data.name ? 1 : -1))
+            .find((r) => r.data.name !== actor.data.data.roleInfo.activeRole);
+          actor.update({ "data.roleInfo.activeRole": newActiveRole.data.name });
+          const warning = `${SystemUtils.Localize("CPR.messages.warnDeleteActiveRole")} ${newActiveRole.data.name}`;
+          SystemUtils.DisplayMessage("warn", warning);
+        } else {
+          actor.update({ "data.roleInfo.activeRole": "" });
+          SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.characterSheet.bottomPane.role.noRolesWarning"));
+        }
+      }
     }
   });
 };